fixed casting

+++ app/Http/Controllers/TestTakes/TestParticipantsController.php
@@ -366,13 +366,13 @@ class TestParticipantsController extends Controller
 
     public function heartbeat($testTakeId, TestParticipant $testParticipant, HeartbeatTestParticipantRequest $request)
     {
-         $testTake = (int)TestTake::whereUUid($testTakeId)->first();
+         $testTake = TestTake::whereUUid($testTakeId)->first();
 
          if ($testTake === null) {
              return Response::make('Failed to process heartbeat of test participant, test take has been deleted.', 500);
          }
 
-        $answer_id = $testTake->getKey();
+        $answer_id = (int) $testTake->getKey();
 
          if ($testParticipant->test_take_id !== $answer_id) {//$testTake->getKey()) {
             return Response::make('Test participant not found', 404);
